+++ integrationTest/test/command/verify/verify.ts
@@ -4,11 +4,11 @@ import * as path from 'path';
 import { readScoreResult } from '../../../helpers';
 
 describe('After running stryker with the command test runner', () => {
-  it('should report 69% mutation score', async () => {
+  it('should report 64% mutation score', async () => {
     const scoreResult = await readScoreResult(path.resolve('reports', 'mutation', 'events'));
     expect(scoreResult.killed).eq(16);
     expect(scoreResult.noCoverage).eq(0);
-    expect(scoreResult.mutationScore).greaterThan(69).and.lessThan(70);
+    expect(scoreResult.mutationScore).eq(64);
   });
 
   it('should write to a log file', async () => {