+++ packages/stryker-api/src/mutant/MutatorFactory.ts
@@ -10,6 +10,18 @@ namespace MutatorFactory {
     constructor() {
       super('mutant-generator');
     }
+
+    /**
+     * Returns the import suggestion for a Mutator
+     * @param name The name of the Mutator the user tried to use.
+     * @returns The name of the package the user may want to install (if it exists).
+     */
+    public importSuggestion(name: string) {
+      if (name === 'typescript') {
+        return `stryker-${name}`;
+      }
+      return `stryker-${name}-mutator`;
+    }
   }
   const mutatorFactoryInstance = new MutatorFactory();