+++ packages/stryker-mocha-runner/src/MochaTestRunner.ts
@@ -116,7 +116,11 @@ export default class MochaTestRunner implements TestRunner {
 
   private additionalRequires() {
     if (this.mochaRunnerOptions.require) {
-      this.mochaRunnerOptions.require.forEach(LibWrapper.require);
+      const modulesToRequire = this.mochaRunnerOptions.require
+        .map(module => {
+          return module.startsWith('.') ? path.resolve(module) : module;
+        });
+      modulesToRequire.forEach(LibWrapper.require);
     }
   }
 }