+++ packages/stryker-html-reporter/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.7"></a>
+## [0.16.7](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.16.6...stryker-html-reporter@0.16.7) (2018-11-13)
+
+
+
+
+**Note:** Version bump only for package stryker-html-reporter
+
+<a name="0.16.6"></a>
+## [0.16.6](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.16.5...stryker-html-reporter@0.16.6) (2018-11-07)
+
+
+
+
+**Note:** Version bump only for package stryker-html-reporter
+
 <a name="0.16.5"></a>
 ## [0.16.5](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.16.4...stryker-html-reporter@0.16.5) (2018-10-15)