+++ packages/stryker-jasmine-runner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/stryker-mutator/stryker/compare/stryker-jasmine-runner@0.2.5...stryker-jasmine-runner@0.2.6) (2018-11-13)
+
+
+
+
+**Note:** Version bump only for package stryker-jasmine-runner
+
 <a name="0.2.5"></a>
 ## [0.2.5](https://github.com/stryker-mutator/stryker/compare/stryker-jasmine-runner@0.2.3...stryker-jasmine-runner@0.2.5) (2018-10-15)