+++ packages/stryker-jest-runner/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.2.7"></a>
+## [1.2.7](https://github.com/stryker-mutator/stryker/compare/stryker-jest-runner@1.2.6...stryker-jest-runner@1.2.7) (2018-11-13)
+
+
+
+
+**Note:** Version bump only for package stryker-jest-runner
+
+<a name="1.2.6"></a>
+## [1.2.6](https://github.com/stryker-mutator/stryker/compare/stryker-jest-runner@1.2.5...stryker-jest-runner@1.2.6) (2018-11-07)
+
+
+
+
+**Note:** Version bump only for package stryker-jest-runner
+
 <a name="1.2.5"></a>
 ## [1.2.5](https://github.com/stryker-mutator/stryker/compare/stryker-jest-runner@1.2.4...stryker-jest-runner@1.2.5) (2018-10-25)