+++ packages/stryker/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.32.1"></a>
+## [0.32.1](https://github.com/stryker-mutator/stryker/compare/stryker@0.32.0...stryker@0.32.1) (2018-11-21)
+
+
+### Bug Fixes
+
+* **log4js:** Don't log log4js category to console ([#1246](https://github.com/stryker-mutator/stryker/issues/1246)) ([479d999](https://github.com/stryker-mutator/stryker/commit/479d999))
+
+
+
+
+<a name="0.32.0"></a>
+# [0.32.0](https://github.com/stryker-mutator/stryker/compare/stryker@0.31.0...stryker@0.32.0) (2018-11-13)
+
+
+### Features
+
+* **error debugging:** add remark to run again with loglevel trace ([#1231](https://github.com/stryker-mutator/stryker/issues/1231)) ([c9e3d97](https://github.com/stryker-mutator/stryker/commit/c9e3d97)), closes [#1205](https://github.com/stryker-mutator/stryker/issues/1205)
+
+
+
+
+<a name="0.31.0"></a>
+# [0.31.0](https://github.com/stryker-mutator/stryker/compare/stryker@0.30.1...stryker@0.31.0) (2018-11-07)
+
+
+### Features
+
+* **clear text reporter:** Prettify the clear-text report ([#1185](https://github.com/stryker-mutator/stryker/issues/1185)) ([a49829b](https://github.com/stryker-mutator/stryker/commit/a49829b))
+
+
+
+
 <a name="0.30.1"></a>
 ## [0.30.1](https://github.com/stryker-mutator/stryker/compare/stryker@0.30.0...stryker@0.30.1) (2018-10-25)